move reset business logic into the reset component
